Adapts to div update using prototype.
I had to add an id to the scrollbox_content div.The id formed by the scrolled div id + "_scrollbox_content". Now I can refresh the div using prototype.
I had to add the appendChild for the up and down buttons in the updateHandle in order to get them redrawed on refresh
Besides, I made the addUpButton and addDownButton, just refactoring.